fix(mqttConfig): Memory stability for multiple calls (#89)
Added a buffer reference to the runCmd so that it knows to free that
buffer after the command was executed.
This caused "Memory leaks" inside the buffer when calling mqttConfig
multiple times.

Added an error log whenever no free buffer is available. There is no
handling for this so this error will likely precede a panic.