fix(WalterModem): MQTT- and HTTP- Receive parser edge case (#125)
fixes #114 
Fixes several scenarios where the MQTT and HTTP payload would be
interpreted as a AT-message rather than binary payload.

Due to the nature of the no-prefix header while receiving data from the
MQTT or HTTP buffer, this is not a fail-free nor "clean" solution like
the socket and coap receive methods.
This can cause unexpected behavior with the parser in the following
scenarios and should be avoided:

`+SQNSNVR` payload should **never** contain:
   * "\r\nOK\r\n"
   * "\r\nERROR\r\n"
   * "\r\n+CME ERROR: "

Result of `AT+SQNSMQTTRCVMESSAGE` - or `AT+SQNHTTPRCV` should **never
start** with:
   * "> " or ">>>"
   * "<<<"
   * "OK\r\n"
   * "ERROR\r\n"
   * "+CME ERROR: "